+++ app/src/main/java/org/android/go/sopt/home/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.home
+package org.android.go.sopt.home.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,22 +7,18 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.ConcatAdapter
-import org.android.go.sopt.R
-import org.android.go.sopt.data.remote.api.ListUsersServicePool
 import org.android.go.sopt.data.remote.model.ResponseListUsersDto
 import org.android.go.sopt.databinding.FragmentHomeBinding
+import org.android.go.sopt.home.TitleViewModel
 import org.android.go.sopt.home.adapter.TitleAdapter
 import org.android.go.sopt.home.adapter.UserAdapter
-import org.android.go.sopt.util.extension.showToast
-import retrofit2.Call
-import retrofit2.Response
 
 class HomeFragment : Fragment() {
     private var _binding: FragmentHomeBinding? = null
     private val binding: FragmentHomeBinding
         get() = requireNotNull(_binding) { "앗 ! _binding이 null이다 !" }
-    private val listUsersService = ListUsersServicePool.listUsersService
-    private val viewModel by viewModels<TitleViewModel>()
+    private val titleViewModel by viewModels<TitleViewModel>()
+    private val viewModel by viewModels<HomeViewModel>()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -35,8 +31,8 @@ class HomeFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        setListUsers()
+        viewModel.getListUsers()
+        getListUserResultObserver()
     }
 
     override fun onDestroyView() {
@@ -48,29 +44,16 @@ class HomeFragment : Fragment() {
         val userAdapter = UserAdapter(requireContext())
         val titleAdapter = TitleAdapter(requireContext())
         userAdapter.submitList(listUsers)
-        titleAdapter.submitList(viewModel.getHomeTitleList())
+        titleAdapter.submitList(titleViewModel.getHomeTitleList())
 
         val concatAdapter = ConcatAdapter(titleAdapter, userAdapter)
 
         binding.rvHomeUsers.adapter = concatAdapter
     }
 
-    private fun setListUsers() {
-        listUsersService.getListUsers().enqueue(object : retrofit2.Callback<ResponseListUsersDto> {
-            override fun onResponse(
-                call: Call<ResponseListUsersDto>,
-                response: Response<ResponseListUsersDto>
-            ) {
-                if (response.isSuccessful) {
-                    response.body()?.data?.let { connectAdapter(it) }
-                }
-            }
-
-            override fun onFailure(call: Call<ResponseListUsersDto>, t: Throwable) {
-                t.message?.let { requireContext().showToast(it) } ?: run {
-                    requireContext().showToast(getString(R.string.server_communication_on_failure))
-                }
-            }
-        })
+    private fun getListUserResultObserver() {
+        viewModel.getListUserResult.observe(viewLifecycleOwner) { listUserResult ->
+            connectAdapter(listUserResult.data)
+        }
     }
 }
